Retire lore_resume (#364)
* Retire lore_resume

Telemetry over 307 sessions recorded zero calls to lore_resume (ADR 0007).
Removes the MCP tool, the `lore resume` CLI verb, and its `--launch`
sub-feature (the cross-integration launcher had no other caller).

- Move the session-note gather helpers (`_iter_session_notes`,
  `_list_wikis`, `_session_date_from_path`) into lore_core/context_pack.py,
  the one live dependency, inlined next to their caller.
- Delete lore_core/resume.py, lore_cli/resume_cmd.py, lore_cli/launcher.py,
  lore_cli/integrations.d/, and the lore:resume skill.
- Unmount `lore resume` from lore_cli/__main__.py; remove the MCP
  registration, handler, and dispatch case from lore_mcp/server.py.
- Sweep remaining references: CONTEXT.md's retrieval-primitive lists,
  skills/context/SKILL.md's fresh-gather pointer, orchestrate-epic's
  working-context bullet, and the two ADR/PRD sentences describing current
  resume-based behavior. Historical CHANGELOG entries and the brainstorm
  analysis that motivated this retirement are left untouched.

Tests: absence checks for both the MCP tool and CLI verb (red before
deletion, green after); a sharded-session-layout regression test ported
into test_context_pack.py to keep the coverage that test_resume.py
carried for the moved code.

Fixes #359

* Clean up stale integrations.d references after lore_resume removal

The launcher's --launch feature was deleted with lore_resume, but
integrations.d/ was still declared as package data and documented as
a live config source.

- pyproject.toml: drop the package-data entry for the deleted dir.
- docs/architecture/config.md: drop the LORE_INTEGRATIONS_DIR env-var
  row and the "install templates" config-source section (renumbered
  the following section).
