+++ lib/lore_cli/__main__.py
@@ -75,7 +75,6 @@ def _build_app() -> typer.Typer:
         migrate_cmd,
         project_cmd,
         quarantine_cmd,
-        resume_cmd,
         scopes_cmd,
         search_cmd,
         session_cmd,
@@ -114,7 +113,6 @@ def _build_app() -> typer.Typer:
     app.add_typer(session_cmd.app, name="session", rich_help_panel=_KN)
     app.add_typer(project_cmd.app, name="project", rich_help_panel=_KN)
     app.add_typer(wiki_cmd.app, name="wiki", rich_help_panel=_KN)
-    app.add_typer(resume_cmd.app, name="resume", rich_help_panel=_KN)
     app.add_typer(lint_cmd.app, name="lint", rich_help_panel=_KN)
     app.add_typer(curator_cmd.app, name="curator", rich_help_panel=_KN)
     app.add_typer(toggle_cmd.on_app, name="on", rich_help_panel=_KN)
