+++ pyproject.toml
@@ -59,7 +59,6 @@ build-backend = "setuptools.build_meta"
 where = ["lib"]
 
 [tool.setuptools.package-data]
-"lore_cli" = ["integrations.d/*.toml"]
 "lore_core" = ["styles/*.md", "styles/vale/*.ini", "styles/vale/WritingRules/*.yml"]
 "lore_core.templates" = ["*.md", "integration-rules/*.md"]
