+++ skills/context/SKILL.md
@@ -22,11 +22,11 @@ Mention `lore doctor` only if the user asks why.
 ## Do not
 
 - Re-narrate the output.
-- Invoke `lore hook session-start` to re-generate — use `/lore:resume`
-  for a fresh gather.
+- Invoke `lore hook session-start` to re-generate — use `lore_search` or
+  `lore_context_pack` for a fresh gather.
 - Read wiki notes, grep the catalog, or call MCP tools.
 
 ## Related
 
-- `/lore:resume` — fresh gather (different from showing the cache)
+- `lore_context_pack` — fresh, scoped gather (different from showing the cache)
 - `lore off` — mute hooks for the session
